refactor(patches): split avatar overlay patch
- Move avatar overlay patching into a dedicated module
- Keep registry and compatibility exports wired to the new module
- Cover stricter i3 compositor hint matching in tests
